refactor(runtime): share agent prompt delay
